refactor(admin): per-screen e2e specs, aria-busy on DataTable, shadcn-faithful use-mobile
Three related cleanups, all flagged in review.

## E2E: one spec per screen

The single `e2e/accessibility.spec.ts` held 5 tests spanning four routes.
Accessibility was a separate concern tested in its own silo. Split into
per-route specs so a11y becomes a property of each screen's contract,
and functional tests (arriving in the editor PR) live alongside their
a11y assertions.

- `e2e/login.spec.ts`, `e2e/bootstrap.spec.ts`, `e2e/dashboard.spec.ts`,
  `e2e/posts.spec.ts` — one per route.
- `e2e/support/rpc-mock.ts` — shared `mockRpc` / `mockSession` helpers
  + `AUTHED_ADMIN` fixture, typed against `AuthSessionOutput`.
- `e2e/support/axe.ts` — `expectNoAxeViolations` helper wrapping the
  WCAG 2.1 AA tag set.
- Old monolithic `accessibility.spec.ts` removed.

## DataTable loading state

`/posts` had empty-state + with-rows axe coverage but no loading-state
coverage. Adds:

- `DataTable` accepts `loadingLabel?: string`; when `isLoading` it renders
  the wrapper with `role="region" aria-label={loadingLabel} aria-busy`.
  Default label is `"Loading"`.
- Posts route passes `loadingLabel="Loading posts"`.
- `posts.spec.ts` `renders skeleton rows while loading` — uses a deferred
  promise to hold `/post/list` pending while axe runs against the
  skeleton, then releases so teardown doesn't hang.

Also drops the unused `unknown` type param on `DataTableProps.columns`
(`ColumnDef<TData>` instead of `ColumnDef<TData, unknown>`).

## use-mobile.ts: back to shadcn default

My earlier `useSyncExternalStore` rewrite was an over-correction — the
shadcn-shipped version works fine and our deviation just creates merge
friction on future `shadcn diff` upgrades. Revert to shadcn's shipped
code + file-level `eslint-disable react-hooks/set-state-in-effect`,
which is the targeted way to exempt the one shadcn-idiomatic pattern
our strict React Compiler lint catches.
