+++ packages/admin/e2e/support/axe.ts
@@ -1,0 +1,13 @@
+import type { Page } from "@playwright/test";
+import AxeBuilder from "@axe-core/playwright";
+import { expect } from "@playwright/test";
+
+// WCAG 2.1 AA coverage — the baseline every admin screen commits to. Call
+// from any spec after the page has rendered its critical content so axe has
+// the DOM in its settled state.
+export async function expectNoAxeViolations(page: Page): Promise<void> {
+  const results = await new AxeBuilder({ page })
+    .withTags(["wcag2a", "wcag2aa", "wcag21a", "wcag21aa"])
+    .analyze();
+  expect(results.violations).toEqual([]);
+}
