+++ packages/admin/src/hooks/use-mobile.ts
@@ -1,27 +1,28 @@
+/* Shadcn-generated hook — ships as part of `shadcn add sidebar`. Kept
+ * verbatim so future `shadcn diff` upgrades don't merge-conflict. The
+ * setState-in-effect pattern is idiomatic for media-query subscription and
+ * functions correctly; we suppress the stricter React Compiler rule only
+ * for this one file.
+ */
+/* eslint-disable react-hooks/set-state-in-effect */
 import * as React from "react";
 
 const MOBILE_BREAKPOINT = 768;
-const QUERY = `(max-width: ${MOBILE_BREAKPOINT - 1}px)`;
 
-// Subscribes to the mobile-breakpoint matchMedia via useSyncExternalStore —
-// no setState-in-effect, no double-render on hydration. SSR returns `false`;
-// real value is read on first client paint.
-function subscribe(callback: () => void): () => void {
-  const mql = window.matchMedia(QUERY);
-  mql.addEventListener("change", callback);
-  return () => {
-    mql.removeEventListener("change", callback);
-  };
-}
+export function useIsMobile() {
+  const [isMobile, setIsMobile] = React.useState<boolean | undefined>(
+    undefined,
+  );
 
-function getSnapshot(): boolean {
-  return window.matchMedia(QUERY).matches;
-}
+  React.useEffect(() => {
+    const mql = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT - 1}px)`);
+    const onChange = () => {
+      setIsMobile(window.innerWidth < MOBILE_BREAKPOINT);
+    };
+    mql.addEventListener("change", onChange);
+    setIsMobile(window.innerWidth < MOBILE_BREAKPOINT);
+    return () => mql.removeEventListener("change", onChange);
+  }, []);
 
-function getServerSnapshot(): boolean {
-  return false;
-}
-
-export function useIsMobile(): boolean {
-  return React.useSyncExternalStore(subscribe, getSnapshot, getServerSnapshot);
+  return !!isMobile;
 }
